refactor: move form examples to separate folder
