+++ examples/app/forms/api/payment-method-registration/route.ts
