+++ examples/app/forms/payment-method-registration/page.tsx
@@ -8,43 +8,51 @@ import { Box, Container, Card, Button, CircularProgress } from '@mui/material';
 import useDataProvider, { useQuery } from '@concepta/react-data-provider';
 
 import {
-  type SignUpFormData,
+  type PaymentMethodRegistrationFormData,
   schema,
   advancedProperties,
   widgets,
+  uiSchema,
 } from './constants';
 
-const uri = '/api/sign-up';
+const uri = '/forms/api/payment-method-registration';
 
-const SignUp = () => {
-  const [formData, setFormData] = useState<SignUpFormData>({
-    firstName: '',
-    lastName: '',
+const PaymentMethodRegistration = () => {
+  const [formData, setFormData] = useState<PaymentMethodRegistrationFormData>({
+    fullName: '',
     email: '',
-    username: '',
-    password: '',
-    passwordConfirmation: '',
-    acceptEmailNewsletter: true,
+    phone: '',
+    address: '',
+    city: null,
+    state: null,
+    zipCode: '',
+    cardNumber: '',
+    expirationDate: '',
+    cvc: '',
+    saveAsDefault: true,
   });
 
   const { post } = useDataProvider();
 
-  const { execute: signUp, isPending: isLoadingSignUp } = useQuery(
+  const { execute: submitPaymentMethod, isPending: isLoadingSubmit } = useQuery(
     (body) => post({ uri, body }),
     false,
   );
 
-  const handleSubmit = async (values: IChangeEvent<SignUpFormData>) => {
-    await signUp(values.formData);
+  const handleSubmit = async (
+    values: IChangeEvent<PaymentMethodRegistrationFormData>,
+  ) => {
+    await submitPaymentMethod(values.formData);
   };
 
   return (
     <Container maxWidth="xs" sx={{ textAlign: 'center', padding: '48px 0' }}>
-      <h1>Sign up</h1>
+      <h1>Payment method registration</h1>
 
       <Card sx={{ marginTop: '48px', padding: '24px' }}>
         <SchemaForm.Form
           schema={schema}
+          uiSchema={uiSchema}
           formData={formData}
           onChange={({ formData }) => {
             setFormData(formData);
@@ -68,10 +76,10 @@ const SignUp = () => {
               disabled={false}
               sx={{ flex: 1 }}
             >
-              {isLoadingSignUp ? (
+              {isLoadingSubmit ? (
                 <CircularProgress sx={{ color: 'white' }} size={24} />
               ) : (
-                'Send'
+                'Submit'
               )}
             </Button>
           </Box>
@@ -81,4 +89,4 @@ const SignUp = () => {
   );
 };
 
-export default SignUp;
+export default PaymentMethodRegistration;
